MB-57078: Reduce the userspace output queue backoff threshold
Reduce the max memory size used before backing off and serve
other clients.

In MB-38007 we saw an increase in rebalance time by using a
max output buffer size of 40MB together with flow control resulting
in a patch setting it to 1MB. This patch aligns the default
setting for the max buffer to that.

The memory used for the output queue may or may not be tracked
by the buckets limit. It'll get tracked if the document size
exceeds 4k (and we don't need to inflate the object). If its
less than 4k the core will allocate that memory and copy the
data into the queue. This means that if you're operating on big
documents you lock up memory inside the engine (causing the
engine to have to evict other data or return tmpfail if it needs
more memory), vs allocating a bunch of memory in the core.

Given that a single tenant may have up to 600 connections even
1MB indicates that a single tenant may use up to 600MB in a
userspace buffer (before writing the data to the kernel). The
entire 1MB buffer needs to be committed to the kernel (and
TCP ACK'd for the kernel to be able to release from SO_SNDBUF
before we can write the next chunk into the kernel).

Note that it is worth mentioning that a normal response packet
is in the ballpark of 30 bytes unless the operation is a
retrieval operation which would then include the requested data.
For clients who use OoO the limit is checked _before_ starting
to execute the commands, which means that the queue may exceed
the threshold once the operations complete. They'll probably
not complete at the exact same time causing some data to be
transmitted into the kernel before the next one completes etc
causing the pipeline to be full.

Note that this may be invalidated when we move to folly with
the support for zero-copy mode, but we would most likely need
to revisit that as part of adding support for zero-copy.

Change-Id: Ic2708c3c9b7fe95301c4513928a05196bb2b97bb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192070
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>