[BP] MB-56644: Make ItemExpel resilient to VBucket rollback
[Backport to 7.1.5]

ItemExpel isn't an atomic operation as that acquires/releases the
CM::lock multiple times. So in the middle of the Expel execution
VBucket rollback might clear the CheckpointList and thus remove the
checkpoint touched by Expel.

Before this fix assertions in CM might trigger and crash the node.
The fix makes ItemExpel resilient to the rollback scenario. ItemExpel
now is capable of identifying the rollback post-conditions and just
gives up if the checkpoint under processing doesn't exist anymore.

Change-Id: I03bf6e82ca9fb799666f5265298e5bcc3ac2b269
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193528
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>