MB-57311: Add support for ClustermapChangeNotificationBrief
ClustermapChangeNotificationBrief push just the version of the
cluster map when it changes instead of the full payload.
It use the same on-the-wire format as ClustermapChangeNotification
expect that the value field is not present.

It intentionally allows the client to ask for both features
as HELO works in the way that the client provides all the features
it would want to enable (or check for), and the server replies
with all of the features that the client asked for that it supports
and enabled for the client. A client who implements both the old
version of the push logic and the new version can ask for both
and when it connects to an old server without support for the
"brief" version the server will enable the old one. If it connects
to a "new" server, the server enables "brief" (and returns that
back to the client).

If we didn't allow for both to be specified a client would have to
run two iterations of HELO when connecting to servers without
support for "Brief" (if they wanted to use the "Full" version).
First with "Brief" and see that the server didn't enable that.
Then run a second HELO specifying the old version

Change-Id: If542ba712f5aa4a6314a510cdd37efa7f40b3cb9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192308
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>