Reject serverless commands in validator
Change the packet validator for serverless commands to reject them
in the validator when executed in "on prem" mode.

Remove the "validation" part of a not-supported packet. Its
framing has already been validated and no one should depend
on "incorrect encoded packet being rejected with another error
message than not supported".

Change-Id: I7763f9f5003cfc817a6c7daece3a002d85bfc06f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192503
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>