MB-56547 [4/3]: Fix OSOBackfillResumesFromNextItemAfterPause test under OSO backfill mode auto
Test 'OSOBackfillResumesFromNextItemAfterPause' was added concurrently
with the OSO backfill mode being implemented; as such it currently
fails as it assumes OSO will always be used if the client supports it,
but that is no longer the case.

Fix by having the test set the OSO mode it requires.

Change-Id: Ibdad7895a1bf92c612ba914c7bc2e09fb13d0c0b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192614
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>