MB-56995: Keep separate counters for important metrics
StoredValues/Blobs/Items can be allocated/freed quickly in short
intervals.

Keep cumulative added/removed counters for those, in addition to the
current value. This ensures we don't "lose" changes to those counts due
to the stat sampling frequency.

Stats for the added/removed counts will be added in a follow up patch.

Change-Id: Ieb4aa2c3b51a9d0fa9b5692ba06f0d8c5d5d088a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192657
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>