MB-56032: Expand FlusherBatchSizeLimitWritersChange with auto-config threads
Add cases for ThreadCount::Default and ThreadCount::DiskIOOptimized;
which require the actual thread count be derived based on CPU cores.

(This is only tangentially related to the linked MB - during
implementaiton of the MB it wasn't clear if the symbolic thread counts
were handled correctly and hence this test was added to confirm).

Change-Id: I4df25280817b547625508689643b81a2cc24c06d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192720
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>