MB-54458: Fix race condition in bucket level stats call
The bucket level stats utilized a task running on the executor
pool to call get_stats in the engine. The logic is a bit
complex as the the underlying engine could dispatch the request
to one of its own threads and return EWB (and signal the
cookie from one of its worker threads), and this is where
the race condition reveal itself. The code (incorrectly)
assumed that the call to the engines get_stats would return
and update the atomic variable _before_ the notification
arrived and the connection was scheduled.

The unit test demonstrated this as ewouldblock_engine don't
dispatch the request to a different thread or do any work before
notifying the cookie. Given that the system is "idle" (no other
active connections) the cookie was run _immediately_ and read
the atomic variable in the task. The state said success (it
wasn't updated with the result from the call to the engine yet)
and an empty set of statistics was sent back to the client.

This patch replace the use of the atomic variable and folly
synchronized vector for the stats with a single folly
synchronized to cover both the error code and the data
generated.

Note that most of the stats calls in the engine is blocking
(the vkey would return ewb if the document needs to be fetched
from disk), and the likelihood of the actual underlying engine
signalling the cookie and the cookie being scheduled to run
before the task release the lock causing the front end thread
to block is really low (in a production environment).

Change-Id: I880da8a091396b073117559c9cb63402065dc585
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192820
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>