MB-57400: Use an estimated items_remaining count for Prometheus stats
Add to CheckpointManager an option to get an estimate of the remaining
items for a cursor - the estimate should provide near constant
performance as it just asks each the checkpoint from curstor to end
for an item count (read of a counter). The max checkpoints is 10, so
the estimated call will only iterate 10 at a maximum. The estimate
will be inaccurate if the cursor is partially through a checkpoint
and will over-count for items already processed.

For the Prometheus requests use the estimated statistic for.

1) The aggregated DCP connection stats (high-cardinality)
2) The aggregated DCP connection stats by type (low-cardinality) but
   only when the  type is !replication.

For cmd_stat, i.e. dcpagg use the accurate mode, so no change from
before this commit.

Change-Id: I712c2a7566735849130092d0f1db07d975cdb2c0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192952
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>