MB-57249: 2/n Modernise core mc_time code
The core of mc_time is a set of functions which keep the memcached
uptime "flowing" and also maintain the "epoch" - which is critical
in calculating expiry.

This change aims to modernise the code whilst minimising any functional
changes, making the code more unit testable is also part of this work.

The main change here is wrapping the code into two new classes (in a
new namespace).

* cb::time::Regulator
* cb::time::UptimeClock

The Regulator is tasked with the job of periodically waking up and
driving the flow of time (our tick)

The UptimeClock owns all of the timekeeping data and code. It provides
a tick function which the Regulator calls. The tick function performs
the same job as the now removed mc_time_clock_tick function,
maintaining the process uptime and watching for system clock "jumps"
(and performing epoch adjustment as per MB-11548).

The change removes lots of static variables from mc_time.cc and
provides improved initialisation and thorough usage of std::chrono.

The code is also adjusted to be unit testable - providing a way to
inject custom clock functions. Note that this method was preferred
over defining clocks via template parameters. With custom clocks, unit
tests can now control the flow of monotonic and system clock time and
inspect how the UptimeClock behaves.

Change-Id: I48fc925d02dbd75de8644223059d518a90ff844b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193302
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>