MB-57644: Don't add 'null' to audit_events.json
For non-production builds we allow audit descriptor files to
be missing, but it resulted in the audit_events.json generated
for ns_server to know about all events contained a "null" entry
instead of a proper module entry.

This patch solves that by skipping all module-definitions
for modules with a "null" json payload.

Change-Id: I55aa64a353cd93b89afba5460bf3f2cfa7263b35
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193194
Reviewed-by: Michael Blow <michael.blow@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>