MB-57249: 8/n Add test coverage for MB-11548
With the flow of time being under control of the unit test add
coverage for MB-11548. This was an issue where if time went
backwards, various time keeping variables became negative and
that triggered incorrect expiry.

The new test covers some of the basics around that MB by changing
time and checking the mc_time functions behave.

Change-Id: Ifddb619d69b7c443831acff757e4fb34ccc08187
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193306
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>