MB-57249: 9/n Change logic of UptimeClock system clock check
This commit resolves MB-57249.

The system clock check exists to monitor if the system clock changes
and in response adjust the epoch (MB-11548). However the check doesn't
work well when the tick callback is irregular.

This commit adjusts the system clock check by comparing system clock
flow against steady clock flow. Ideally they are both ticking forwards
equally, and only if the system clock does not follow the steady clock
do we need to warn and adjust the epoch.

For example is 100 seconds of steady time elapse between checks, expect
that 100s of system time also has elapsed (with some small tolerance).
If the system clock has not progressed as expected, warn and adjust the
epoch.

Change-Id: Ic2bc719d382b4e6d28a2c0c95632397124de2b17
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193307
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>