MB-57711: Fix calculation of num_auxio_threads=default
+Problem+

The AuxIO thread settings API incorrectly interprets "default" if it
is used to set the thread count after initial startup - it sets the
thread count to 0 (i.e. no AuxIO jobs can run), instead of calculating
based on the CPU count.

(Note NonIO threads have the same issue, they will be fixed in a
subsequent patch).

+Background+

The issue is in how "zero" as a thread count is parsed - different
parts of the system try to use zero for different purposes - either to
mean "no threads" (unit tests & TAF), or to mean "automatically
select".

Specifically, for Reader / Writer threads, a count of "0" historically
meant "auto-select based on CPU count". When we added reader / writer
thread counts into the REST / UI in 6.5, there were 3 values which
could be passed from ns_server to kv for readers & writers:

- "default" - Use automatically calculated values from pre 6.5
  (generally based on CPU count).

- "disk_io_optimized" - Use automatically calculated values with more
  aggressive settings.

- A positive integer - explicitly use that many threads.

Given "0" already meant "auto-select", KV-Engine actually mapped the
string "default" from ns_server to the value "0" for reader/writer
threads.

Note at this point, AuxIO (and NonIO) threads couldn't be configured
via ns_server; only via epctl as they had historically been done. When
we did add support for configuring AuxIO / NonIO (MB-49977), the same
"default" value was used to indicate that auto-selected thread count
was used. However, that effectively was also mapped to zero.

When the ExecutorPool is initially constructed this is fine, as the
number of AuxIO threads were derived, and zero meant "auto-configure"
- i.e. calcNumAuxIO() is called on the input value.

However if a later call was make to change the config (to "default"),
this was again mapped to zero, and the number of threads directly set
to that value (without calling calcNumAuxIO).

+Solution+

Introduce a new enum type - ThreadPoolConfig::AuxIoThreadCount - which
is used as the type for AuxIO thread counts. This defines a "default"
enum value as "-1", meaning that "0" is still available as a valid
value (for test uses at least).

Change-Id: Icf91d7ea0dd385611636b7ea88d5094d19129bec
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193578
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>