MB-57754 Release Magma KeyIterator when scan is yielded
Magma owns the value buffer internally when supplying values to KV as
part of KeyIterator::GetValue. Once a value has been loaded into memory,
it is only released upon iterator destruction or issuance of a iterator
Next() call.

Right now KV can snooze a backfill without desturcting the iterator.
This causes the memory to be held by the iterator. If KV is round
robining across multiple backfills, this can happens for many backfills
and bloating the memory usage. This could result in none of the
backfills proceeding if we've exceeded the thresholds.

Fix this by creating the iterator upon every backfill resume and
destroying it upon every pause.

This is similar to change we made as part of MB-52668 for SeqIterator.

Change-Id: I06139076195675ae3a1c5b73afaf18c38a97cf23
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193560
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Rohan Suri <rohan.suri@couchbase.com>