MB-57772: Remove DCPBackfill object when ActiveStream destructs
Prior to this commit, the removal (and deallocation of memory) of
a DCPBackfill object relied on the backfill manager running the
backfill and finding the backfill reports a "failure" (when it
cannot weak_ptr::lock the ActiveStream). This approach is flawed
as there is no guarantee that the backfill manager will execute
that code path, e.g. when memory usage is high.

As the MB demonstrated, it's possible for a DCP client to live
lock KV-engine by cycling through stream requests and growing the
backfill queue to the point where the queue itself is the largest
user of memory.

This fix adds a deallocation path directly from ~ActiveStream so
that deallocation of a DCPBackfill object occurs serially with the
destruction of the owning ActiveStream.

The destruction works by having each scheduled DCPBackfill assigned
a UID (a read of a static monotonic u64). The ActiveStream keeps a
copy of the UID and uses it in a removeBackfill call when
destructing.

Change-Id: Iae4947e4dc3838ba6b25682d5758154ea5cd69ef
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193640
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>