MB-57817: Add Item "history" setting to Magma metadata
Add a new version of magmakv::MetaData, V2.

V2 first extends the metadata so that a history bool can be stored.
The history bool goes into the existing V0 bit field, but is only
usable when version is V2.

Next V2 allows for optional addition of the durability data. This
works on the encode path by adding a trip wire "bool" to the set
methods which add any V1 durability data. If the durability data
is set, then the structure encodes as if it were V1.

On the decode side, V2 now optionally looks for trailing data after
the V0 decode, if the input buffer has more data it's assumed to be
the V1 data.

Change-Id: Ibc2145569f2e04e911da8c33d6abf07b62325290
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193805
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>