MB-57864: [subdoc] Allow operations on multiple xattr keys
Allow a single subdoc operation to operate on different xattr keys.

In the initial specification for subdoc this was not allowed, as we
wanted to be able to store the various xattr keys in separate locations.
To simplify the implementation of the initial version of XATTRs we
ended up storing the xattrs together with the document:

    +---------------------------------------------+
    | 4 bytes indicating the entire xattr segment |
    +---------------------------------------------+
    | N bytes containing encoded xattrs           |
    +---------------------------------------------+
    | Y bytes containing the document value       |
    +---------------------------------------------|

The xattrs are stored as a sequence of key-value pairs with
a length field to allow for easy traversal without having
to parse each individual byte:

    +---------------------------------------------|
    | 4 byte length field of the entire key value |
    +---------------------------------------------|
    | zero terminated key                         |
    +---------------------------------------------|
    | zero terminated value                       |
    +---------------------------------------------|

In retrospect we should probably have stored the xattr section
as a blob of JSON as that would have allowed us to use the xattr
section directly with the sub-json code instead of building
up a JSON document of the xattr kv-pair for every operation
(and this patch change that to _all_ kv-pairs).

It could be that we should introduce a new datatype field to
represent JSON encoded xattrs to get rid of encoding/decoding,
but that needs to be implemented separately as that is a big task
involving having to rewrite documents as part of DCP to clients
without support for the new format etc.

Change-Id: I2e6680fe1fa2d011126cd9873571e7615c215b9e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193928
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>