[testapp] Don't restart the prometheus endpoint for every reload
The unit tests pass 0 for the port number to use for prometheus
server to let the kernel pick an available port number to use.
The code in settings reload compared the provided port number
with the current port number (0 to nnn) to determine if the
configuration changed; --> it was different every time causing
the prometheus endpoint to be changed every time we tried
to reload the settings.

Change-Id: I19d02bc8520b8070befbf05fedfcd8efe069075c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194054
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>