MB-56648: [NIM] Run operations on server array in nim-thread
Run all of the operations which operate on the vector of server
sockets in the same thread to avoid potential race conditions.

Internally libevent use a mutex to protect the event structure
registered in libevent (via add/del event), but the object could
be used by the other thread (in a callback).

To work around race conditions use runInEventBaseAndWait to
serialize all access to the server sockets. Note that this
is _not_ a hot path in our system (only used when the user
tries to change/list the interface descriptions).

Change-Id: Iad40dec29e2a35ac5be2449f1b02d63832dc9918
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194126
Well-Formed: Restriction Checker
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>