MB-46740: Decrease dcp_backfill_byte_limit from 20MB to 2MB
The param enforces a limit (bytes) on the allocation that the single
DCP Producer connection can make into its streams readyQs (aggregated
for all the streams).

Note that we always allow at least one backfilled item into the readyQ,
so the new default value comes with a possible error in the case of
large documents. Apart from that exception, the new value is expected
to cut the max possible allocations (by backfill) in the readyQ to 1/10
compared to the previous value.

Change-Id: I5fd456f8099429719c4aebf9ca1afef82b682ac0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194067
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>