MB-57979: The retry-counter should be stored in the context
Subdoc allows "automatic retries" for mutations if the client
use the CAS wildcard (0) to solve concurrency "problems" when
another client replaced the document while this client performed
its mutations:

    while (#retries < max_retries) {
       current_doc = engine-get();
       execute subdoc operation
       try to replace new doc by using current cas
    }

The problem was however that the code didn't handle
that interacting with the underlying engine could block, and
reset the retry counter every time.

All the state variables for a single pass for the operation
was stored in a class called SubdocCmdContext. It is a nice
separation as we could simply swap out the entire object and
start over, but at the same time we don't have anywhere to
store the retry counter.

This patch introduce SubdocumentCommandContext, but do
not rename the SubdocCmdContext to limit the size of the patch.
It may (should) be performed in a series of patches.

Change-Id: I8d07f4e8a7d2417dfcbf7cbef182ab863c164286
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194254
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>