MB-58032: Remove support for pre-5.0 DCP Producers
As part of MB-29441 (added in 5.1.1) we added support in ep-engine's
DcpConsumer to detect if the Producer node supports a reduced no-op
interval (added in 5.0.0); if not (i.e. in a mixed-mode cluster during
upgrade) we fallback to the previous behaviour.

We do not support running in a mixed-mode with more than one major
version difference - so as of 7.x we do not actually support any 5.x
version. As such, we can remove the code to handle both cases and
simply refuse DCP between a 7.x and <5.0.0 version.

(We could go further and refuse to connect to anything pre-6.0;
however I haven't looked into if we have a reliable way to detect <6.0
- and the motivation here is to remove the pre-5.0 specific code to
simplify future DCP noop changes.)

Change-Id: I26fcf5cf551e5f49d293a4cb3b0891e5784df545
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194663
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>