MB-55754: Back off the CPU after executing commands for 25ms
Previously each execution of a cookie allowed up for a certain
number of cookies to be "executed" until yielding to allow
the next connection to run.

This patch extend the logic to only allow a connection to
execute cookies until its timeslice exceeds. Please note
that exectuion of a cookie still may exceed the timeslice
as the timeout is checked _before_ accepting / calling execute
on a cookie.

The executor for Subdoc commands has been updated to check
for the timeout (and possibly yield) for retry attempt.

Change-Id: If59ebcebae3759c0317ff147914ae0de1f21ab52
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194205
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>