MB-58075: Fix intermittent failure in BucketQuotaChangeTest.QuotaChangeUp
As seen during CV for an unrelated change:

    [ RUN      ] EphemeralOrPersistent/BucketQuotaChangeTest.QuotaChangeUp/magma_persistent_full_eviction
    unknown file: Failure
    C++ exception with description "CheckedExecutor::runCurrentTask(): Expected task: "Changing bucket quota" got "Paging out items."" thrown in the test body.
    [  FAILED  ] EphemeralOrPersistent/BucketQuotaChangeTest.QuotaChangeUp/magma_persistent_full_eviction, where GetParam() = "backend=magma:bucket_type=persistent:item_eviction_policy=full_eviction" (8984 ms)

This appears to be due to the test taking greater than the default
ItemPager frequency (5s) to attempt to run the "next" NonIO task -
which is expected to be BucketQuotaChangeTask.

It's not clear why there's such a delay in this test; however, we do
see similar runtimes with the subsequent Magma task so perhaps some
IO-related issue in setting up Magma?

Avoid the issue by increasing the ItemPager sleep time in this test to
an arbitrarily large value (100s).

Change-Id: I1e538ba93c038e3653eaf9bdfb63eedba709623c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194664
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>