MB-57325 [1/2]: Allow writer threads to be proportional to nCPUs
+Background+

The number of threads created for 'Disk I/O Optimized'
readers/writers was equal to the number of CPUs. However, in MB-56202 we
observed that the total available GCP disk IOPS could not be achieved
without increasing the IO threads to 2x the number of CPU cores in the
test. Therefore, we would like to increase the total number of
reader/writer threads proportionaly to the total number of CPU cores.

+Change+

This patch changes the way the writer thread pool size is calculated to
spawn num_io_threads_per_core * nCPUs, instead of one thread per CPU
core.

Change-Id: I362eb0353ee36c024c43d86831c583692df3caeb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/196733
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>