MB-57325 [2/2]: Allow reader threads to be proportional to nCPUs
+Background+

The number of threads created for 'Disk I/O Optimized'
readers/writers was equal to the number of CPUs. However, in MB-56202 we
observed that the total available GCP disk IOPS could not be achieved
without increasing the IO threads to 2x the number of CPU cores in the
test. Therefore, we would like to increase the total number of
reader/writer threads proportionaly to the total number of CPU cores.

+Change+

This patch changes the way the reader thread pool size is calculated to
spawn num_io_threads_per_core * nCPUs, instead of one thread per CPU
core.

Change-Id: Ib648a61f1cf9e0d1fdf2f6584165a508ece13e90
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/196739
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>