MB-54032: Copy Manifest into engine specific to avoid shared ownership
Previously a raw pointer was stored in engine specific storage from
which a unique_ptr was constructed. Attempting to use a
shared_ptr/weak_ptr combination revealed that the shared_ptr could
be destroyed before a weak_ptr access.

By copying the Manifest into engine specific (owned by cookie) and
task, we avoid shared ownership, simplifying memory management.

Copying the manifest changed the order in which collections are
created in a test case, which is not guaranteed due to the use of
unordered_map. The test was modified to not check this.

The patch also fixes the assignment operator of Manifest to correctly
reset the uid even if it breaks monotonicity. A test was added to
check this behaviour and constructor correctness.

Change-Id: I57bdb946ccc195c4bccfc56df9b023bd2bc4452f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195108
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>