MB-39344: Ensure no logic change in CM::getSnapshotInfo()
The logic of the function is conditional on whether the open checkpoint
contains non-meta items or not.

Given that by MB-39344 now checkpoints can be made empty by ItemExpel
(ie, after being non-empty), and given that is a new state, then in the
function we want to enforce the same logic as prior MB-39344.

I've added unit test coverage for CM::getSnapshotInfo() in this patch.
Actually that proves that the function behaviour was already correct
before the small prodution adjustement that I make here. See comments
in the code for details.

Change-Id: Ib1f957e44f1f128fb9f9aeeb7503b456681773cc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195181
Well-Formed: Restriction Checker
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>