MB-39344: Ensure no logic change in CM::getVisibleSnapshotEndSeqno()
The logic of the function is conditional on whether the open checkpoint
contains non-meta items or not.

Given that by MB-39344 now checkpoints can be made empty by ItemExpel
(ie, after being non-empty), and given that is a new state, then in the
function we want to enforce the same logic as prior MB-39344.

Change-Id: Ib2380d3e069ed99a4f19d5d89a306ea3bcea2a73
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195255
Well-Formed: Restriction Checker
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>