MB-58091: Change the default Metered from true to false
The original metering change added "metered" defaulted to true. This
was to work around the fact that at the time changing collection meta
data was a PITA, so we just assume that the default collection has
metering and serverless "just works".

With changes from CDC which added collection modifications this
can be undone (resolving the MB), metering is defaulted to false and
it can be later enabled by ns_server by explicitly adding metered:true
to a collection entry.

Notes:

Metering against the default collection breaks in this commit (to be
fixed). This occurs because a number of collection lookup paths use
a "static" result when the collection is the default - resulting in
metering = false.

Additionally, with this commit enabling metering on the default
collection will fail as it is a change of a immutable property.
A second commit will allow for metering to change.

Change-Id: I83c6a0c7bb3c632d7b5aaca84d0522aa73cdda27
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195270
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>