Don't create two connections in the CLI tools
McProgramOptions incorrectly asked clone() to connect to the
server and copy over the enabled features. This caused the log
to look like:

    INFO 59: HELO nothing
    INFO 59: Client authenticated as <ud>Administrator</ud>
    INFO 59: HELO XERROR <ud>Administrator</ud>

In addition to that there was a socket reconnect before the
line with client authenticated (caused by the reconnect).

Given that we don't have any features to enable we should
delay connect and helo until we've successfully authenticated.

Change-Id: I9b0c30e3dc4f3f67feecef45d6efd88097d33333
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195314
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>