MB-57767: Relax the invariant on nextSnapStart (non-monotonic)
The CheckpointManager can contain checkpoints which have non-monotonic
snapshot start sequence numbers. This can happen on a replica if a
checkpoint containing only meta items is created. We can end up creating
such a checkpoint if a set_vbucket_state{replica} is in its own
checkpoint and we start streaming from the active just after the
set_vbucket_state is queued into the CM.

ActiveStream::processItems tracks a nextSnapSeqno, which it updates as
it sees checkpoints from the CM. The nextSnapSenqo is Monotonic, however
in the case described, it happens to not be. The meta-only checkpoint
updates the nextSnapStart seqno, but we never send a snapshot for that
checkpoint.

Scenario:

Node A, currently active, processes the following sequence of mutations,
ending in set_vbucket_state, changing the state to replica.

    0         1      2       3         3                3                      3
[ckpt_start, m(x), m(y), ckpt_end] [ckpt_start, set_vbucket_state(replica), ckpt_end]

Node A, now a replica, sets up a replication stream to Node B, the new
active. The stream request is [2, inf], as after 2 we have all meta
items.

Node B (active) sees:

           3                        3       4      5      6
[set_vbucket_state(active)],  [ckpt_start, m(w), m(x), chkp_end]
               ^
      meta (not sent)

Because of MB-50333, B sends a Checkpoint[seqno:{3, 6}, snap:{2, 6}].
That is, the snapshot startSeqno is set to the requested startSeqno,
even though the first mutation we send is seqno 3.

Now A has the following checkpoints.

   0           1    2      3            3               3                       3         3          3     4      5
[ckpt_start, m(x), m(y), ckpt_end] [ckpt_start, set_vbucket_state(replica), ckpt_end] [ckpt_start, m(w), m(x), chkp_end]
     snap: {1, 3}                           snap: {3, 3}                                    snap:{2, 6}

Note that in the above, the snapStart goes back. Later, node A flips
back to active. As we process items for streaming, we hit the
invariant failure, because we see the next snapStart go back.

Detecting this edge case (meta-only checkpoint -> do not set
nextSnapStart) is tricky, as we process items in a foward-only fashion.

We do however, already have a monotonic lastSentSnapStartSeqno, which is
set when we create the SnapshotMarker, guarding against the behaviour
the monotonic check on nextSnapSeqno was supposed to guard against.

Remove the Monotonic from nextSnapStart, lastSentSnapStartSeqno already
checks that we don't go back for the SnapshotMarkers that we send.

Change-Id: Ica0831adae5688801e913fd4a155f5ef7e617a32
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195341
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>