Fix ImposedUsersMayMeter test to read the collection
The test was incorrectly writing to a collection but always reading
from the default collection, so never testing the unmetered by
collection case.

The test now expects that for UnmeteredByCollection that the
impersonated operation doesn't get metered. In a real deployment only
the system collections are unmetered in this way, and would be subject
to further access controls compared to a regular client.

Change-Id: Iff8fd3b1cfb82b1e1bc2f833dd40029da101fba0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195354
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>