MB-58261: tryBackfill must be true when checkpoints are empty by expel
In the MB a startSeqno of 0 was passed to registerCursorBySeqno whilst
the first checkpoint was "empty by expel".

This leads to registerCursorBySeqno finding the lowest available seqno
but not returning tryBackfill = true - meaning that DCP skips the
backfill and produces and incorrect replica. This occurs because
when a checkpoint is "empty by expel" it returns 0 as the minimum
seqno (as it does not know the minimum seqno), which causes the
registerCursorBySeqno code to enter a path leading to this issue.

This commit checks for the emptyByExpel case and enables tryBackfill

Change-Id: I8c7b4ca61fa6ddfdaf7bb277f28925aac07138ed
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195382
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>