MB-58154: Early close of TLS connections
ns_server never got around to use ifconfig to define the
network interfaces and continues to provide the interfaces
in memcached.json during startup and then later on send
the tls configuration (certificates etc) via the ifconfig
command. (MB-39620)

Work around this by checking if TLS is configured as part
of accepting a connection on an interface marked for TLS
and close the connection with a message saying that TLS
isn't configured (with a link to this MB).

Change-Id: Iad2118407d92afdaeb04b3cfcfde1acf34a0528b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195645
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>