MB-56973 [6/6]: Support sub-second DCP noop interval on Producer
Add support for negotiating a sub-second 'set_noop_interval' to
DcpProducer - up to the ConnManager interval which is 0.1s.

This means that if a Dcp Consumer - such as KV-Engine's own
replication consumer - requests sub-second DCP noops; the producer
will honor the request instead of rejecting it.

To support this we need to update the DcpProducer to track last send /
receive time at sub-second granularity. This is done by switching to a
chrono-style time_point returned by the recently added
UptimeClock::now() method, via the ServerCoreIface.

With this patch, DCP NOOPs on KV-Engine replication streams will be
exchanged every 100ms by default (assuming Trinity producer &
consumer).

Change-Id: I8d45f2cea790c7ae22c66067666dfe1fd50a2a9d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195729
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>